% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/feelings.R
\name{feelings}
\alias{feelings}
\title{Feelings}
\usage{
feelings(.data)
}
\arguments{
\item{.data}{tibble containing at the minimum a column called "word"}
}
\value{
tidy tibble with a row for each word/sentiment pair from NRC/.data
}
\description{
Feelings
}
