% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data-sample_speech_data.R
\docType{data}
\name{sample_speech_data}
\alias{sample_speech_data}
\title{Sample Speech Text Data}
\format{
An object of class \code{tbl_df} (inherits from \code{tbl}, \code{data.frame}) with 11 rows and 3 columns.
}
\usage{
sample_speech_data
}
\description{
Text of speeches by prominent figures
}
\details{
Sources include:
\itemize{
\item kinginstitute.stanford.edu
\item er.jsc.nasa.gov
\item ourdocuments.gov
\item americanrhetoric.com
\item jfklibrary.org
\item teachingamericanhistory.org
\item presidency.ucsb.edu
}
}
\keyword{datasets}
