% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/scrape_pdf.R
\name{scrape_pdf}
\alias{scrape_pdf}
\title{Scrape pdf for text}
\usage{
scrape_pdf(.data)
}
\description{
Scrape the pages of text from pdfs containing Supreme Court opinions;
cleans text and organizes relevant metadata
}
