% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/spch_tkn_map.R
\name{spch_tkn_map}
\alias{spch_tkn_map}
\title{Create Tibble from Several Speech Sources}
\usage{
spch_tkn_map(source_list)
}
\arguments{
\item{source_list}{}
}
\value{

}
\description{
spch_tkn_map receives a tibble (source_list) containing character vectors for url,
speech title, and speech author for use with the speech_tokens function.
}
\details{
purrr::pmap_dfr is called on source_list, mapping each row in source list
as the inputs to speech_tokens(url, title, author).

spch_tkn_map returns a tokenized tibble of all words in each speech, with
columns reflecting title and author
}
