% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data-spchlist.R
\docType{data}
\name{spchlist}
\alias{spchlist}
\title{Speech list}
\format{
An object of class \code{tbl_df} (inherits from \code{tbl}, \code{data.frame}) with 11 rows and 4 columns.
}
\usage{
spchlist
}
\description{
List of local html files containing text of speeches by prominent public
figures.
}
\details{
Sources include:
\itemize{
\item kinginstitute.stanford.edu
\item er.jsc.nasa.gov
\item ourdocuments.gov
\item americanrhetoric.com
\item jfklibrary.org
\item teachingamericanhistory.org
\item presidency.ucsb.edu
}
}
\keyword{datasets}
