% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/speechscrape.R
\name{speechscrape}
\alias{speechscrape}
\title{Speech Scrape}
\usage{
speechscrape(url, slctr = "p+ p")
}
\arguments{
\item{url}{string containing location of chosen html file}

\item{slctr}{string specifying css selector for chosen text}
}
\value{

}
\description{
Here I am describing some things.
}
