% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/scrape_pdf.R
\name{scrape_pdf}
\alias{scrape_pdf}
\title{Scrape pdf for text}
\usage{
scrape_pdf(.data)
}
\arguments{
\item{.data}{Text vector where each element represents a page from a pdf}
}
\value{
Tibble containing cleaned text, case name, opinion type, and author
}
\description{
Scrape the pages of text from pdfs containing Supreme Court opinions;
cleans text and organizes relevant metadata
}
