% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/speech_tokens.R
\name{speech_tokens}
\alias{speech_tokens}
\title{Speech Tokenizer}
\usage{
speech_tokens(text, title, author)
}
\arguments{
\item{text}{string containing full text to be tokenized}

\item{title}{string value of speech title}

\item{author}{string value of speech author's full name}
}
\value{

}
\description{
speech_tokens function receives a url input and uses CSS selectors to parse
and tokenize the portion of the target url believed to contain the txt of a
speech; returns a tibble containing each word in that speech
}
