% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/wordcount_plot.R
\name{wordcount_plot}
\alias{wordcount_plot}
\title{Plot Word Count}
\usage{
wordcount_plot(speech_tkn)
}
\arguments{
\item{speech_tkn}{tokenized speech object input}
}
\value{

}
\description{
The speech_tkn function receives as input a tidy tokenized tibble of text
from a speech, counts the number of occurences of each word
}
\examples{
url <- "https://kinginstitute.stanford.edu/king-papers/documents/i-have-dream-address-delivered-march-washington-jobs-and-freedom"
title <- "I Have a Dream"
author <- "Dr. Martin Luther King, Jr."
wordcount_plot(speech_tokens(url, title, author))
}
